refactor: separate order status icons from pure constants
Move ORDER_STATUS_ICONS to lib/constants/order-icons.ts so that
lib/constants/orders.ts has no UI imports (@hugeicons). This keeps
the tested module free of unnecessary dependencies.
